:technologist: Chores: add helper functions to return resource error when http, io exception occured
