+++ authentication/authentication-data/src/main/java/com/sowhat/authentication_data/di/AuthModule.kt
@@ -1,0 +1,4 @@
+package com.sowhat.authentication_data.di
+
+object AuthModule {
+}
