+++ authentication/authentication-data/src/main/java/com/sowhat/authentication_data/mapper/AuthMapper.kt
@@ -1,0 +1,2 @@
+package com.sowhat.authentication_data.mapper
+
