+++ authentication/authentication-data/src/main/java/com/sowhat/authentication_data/repository/AuthRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package com.sowhat.authentication_data.repository
+
+class AuthRepositoryImpl {
+}
