+++ authentication/authentication-domain/src/main/java/com/sowhat/authentication_domain/model/NewMemberInfo.kt
@@ -1,0 +1,3 @@
+package com.sowhat.authentication_domain.model
+
+data class NewMemberInfo()
