+++ authentication/authentication-domain/src/main/java/com/sowhat/authentication_domain/repository/AuthRepository.kt
@@ -1,0 +1,4 @@
+package com.sowhat.authentication_domain.repository
+
+interface AuthRepository {
+}
