+++ core/common/src/main/java/com/sowhat/common/wrapper/Resource.kt
@@ -1,0 +1,3 @@
+package com.sowhat.common.wrapper
+
+sealed class Resource
